requirements.txt
streamlit
openai
faiss-cpu
numpy
tiktoken
pdfplumber
python-dotenv
langchain
